Configure game screens switching with 10 mock questions
